requirements.txt
allure-behave
allure-python-commons
attrs==20.2.0
behave==1.2.6
iniconfig==1.1.1
packaging==20.4
parse==1.18.0
parse-type==0.5.2
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest
selenium~=4.2.0
websocket-client~=1.3.2
requests~=2.28.0
webdriver-manager
behavex
JPype1